fix int->long longに修正

+++ Math/divisor.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 vector<long long> divisor(long long n) {
     vector<long long> ret;
-    for (int i = 1; i*i <= n; ++i) {
+    for (long long i = 1; i * i <= n; ++i) {
         if (n % i == 0) {
             ret.push_back(i);
             if (i * i != n) ret.push_back(n / i);
